requirements.txt
Django
djangorestframework
bs4
lxml
django-cors-headers
